moved all of image & texture to hpp
with code in cpp file, in separate translation
unit, operator[] wasn't inlined (I checked asm output).
therefore this made it possible to inline the hot
path (FillScanline)
result fps went ~80 -> ~110
