+++ CMakeLists.txt
@@ -18,7 +18,6 @@ add_executable(main "src/main.cpp"
     "src/graphics/sdl_wrap.cpp"
     "src/scene/mesh.cpp"
     "src/app.cpp"
-    "src/graphics/image.cpp"
     "src/util/timer.cpp"
     "src/scene/camera.cpp"
     "src/view.cpp"
