+++ src/fs/texture_reader.cpp
@@ -21,49 +21,26 @@ struct STBDeleter {
 namespace impl {
 
 struct Texture {
-  Texture(Width width, Height height, std::unique_ptr<unsigned char, STBDeleter>&& ptr)
-      : width_(width), height_(height), data_(std::move(ptr)) {
+  Texture(std::unique_ptr<unsigned char, STBDeleter>&& ptr) : data_(std::move(ptr)) {
   }
 
-  Width width_;
-  Height height_;
   std::unique_ptr<unsigned char, STBDeleter> data_;
 };
 
 }  // namespace impl
 
-Texture::Texture(std::unique_ptr<impl::Texture>&& impl) : impl_(std::move(impl)) {
+Texture::Texture(Width width, Height height, std::unique_ptr<impl::Texture>&& impl)
+    : impl_(std::move(impl)), width_(width), height_(height), buf_(impl_->data_.get()) {
 }
 
-Texture::Texture(Texture&& other) noexcept : impl_(std::exchange(other.impl_, nullptr)) {
+Texture::Texture(Texture&& other) noexcept
+    : impl_(std::exchange(other.impl_, nullptr)),
+      width_(other.width_),
+      height_(other.height_),
+      buf_(std::exchange(other.buf_, nullptr)) {
 }
 
 Texture::~Texture() {
-}
-
-Pixel Texture::operator[](Index ind) const {
-  size_t pix_ind = ind.row_ * GetWidth() + ind.col_;
-  Pixel pixel;
-  pixel.r_ = GetBuffer()[pix_ind * 3 + 0];
-  pixel.g_ = GetBuffer()[pix_ind * 3 + 1];
-  pixel.b_ = GetBuffer()[pix_ind * 3 + 2];
-  return pixel;
-}
-
-Width Texture::GetWidth() const {
-  return impl_->width_;
-}
-
-Height Texture::GetHeight() const {
-  return impl_->height_;
-}
-
-const unsigned char* Texture::GetBuffer() const {
-  return impl_->data_.get();
-}
-
-unsigned char* Texture::GetBuffer() {
-  return impl_->data_.get();
 }
 
 TextureReader::~TextureReader() {
@@ -80,9 +57,8 @@ Texture TextureReader::LoadImage(std::filesystem::path path) {
     LOG(FATAL) << "Could not load texture from an image";
   }
   auto buf_wrapped = std::unique_ptr<unsigned char, STBDeleter>(buf, STBDeleter{});
-  auto impl_tex =
-      std::make_unique<impl::Texture>(Width(width), Height(height), std::move(buf_wrapped));
-  return {std::move(impl_tex)};
+  auto impl_tex = std::make_unique<impl::Texture>(std::move(buf_wrapped));
+  return {Width(width), Height(height), std::move(impl_tex)};
 }
 
 }  // namespace renderer
