+++ src/fs/texture_reader.hpp
@@ -15,19 +15,37 @@ struct Texture;
 
 class Texture {
  public:
-  Texture(std::unique_ptr<impl::Texture>&& impl);
+  Texture(Width width, Height height, std::unique_ptr<impl::Texture>&& impl);
   Texture(Texture&& other) noexcept;
   ~Texture();
 
-  Pixel operator[](Index ind) const;
+  Pixel operator[](Index ind) const {
+    size_t pix_ind = ind.row_ * GetWidth() + ind.col_;
+    Pixel pixel;
+    pixel.r_ = GetBuffer()[pix_ind * 3 + 0];
+    pixel.g_ = GetBuffer()[pix_ind * 3 + 1];
+    pixel.b_ = GetBuffer()[pix_ind * 3 + 2];
+    return pixel;
+  }
 
-  Width GetWidth() const;
-  Height GetHeight() const;
-  const unsigned char* GetBuffer() const;
-  unsigned char* GetBuffer();
+  Width GetWidth() const {
+    return width_;
+  }
+  Height GetHeight() const {
+    return height_;
+  }
+  const unsigned char* GetBuffer() const {
+    return buf_;
+  }
+  unsigned char* GetBuffer() {
+    return buf_;
+  }
 
  private:
   std::unique_ptr<impl::Texture> impl_;
+  Width width_;
+  Height height_;
+  unsigned char* buf_;
 };
 
 class TextureReader {
