+++ src/util/parallelism.hpp
@@ -12,7 +12,7 @@ template <typename T>
 std::unique_ptr<T[]> FillParallel(size_t sz, T elem) {
   std::unique_ptr<T[]> result = std::make_unique_for_overwrite<T[]>(sz);
   T* begin = &result[0];
-  std::fill(std::execution::par, begin, begin + sz, elem);
+  std::fill(std::execution::par_unseq, begin, begin + sz, elem);
   return result;
 }
